backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
stripe==10.12.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
httpx==0.27.2
slowapi==0.1.9
langchain==0.3.7
langchain-community==0.3.7
faiss-cpu==1.9.0
sentence-transformers==3.3.1

# Dev/test dependencies
pytest==8.3.3
pytest-asyncio==0.24.0
